qnd: should not delete the reload.base file from CLI - in machines running both batch and API, we want the reload to be performed by the apache
git-svn-id: svn+ssh://kelev.kaltura.com/usr/local/kalsource/backend/server/tags/gemini_2012_10_22@89730 6b8eccd3-e8c5-4e7d-8186-e12b5326b719

+++ server_infra/kConf.php
@@ -79,7 +79,7 @@ class kConf extends kEnvironment
 		if(function_exists('apc_store'))
 		{			
 			$res = apc_store(self::APC_CACHE_MAP, self::$map);
-			if($reloadFileExists && $res)
+			if($reloadFileExists && $res && PHP_SAPI != 'cli')
 			{
 				$deleted = @unlink("$cacheDir/base.reload");
 				error_log("Base configuration reloaded");
